# services/api/docs/swagger.yaml
basePath: /api
definitions:
...
  home.CPU:
    properties:
      avg_percent:
        type: number
      total:
        type: integer
      used_units:
        type: number
    type: object
  home.CurrentStats:
    properties:
      _Last stats reset:
        type: string
      Authentication failures:
        type: integer
      Average auth time:
        type: string
      Average session time:
        type: string
      Closed due to error sessions:
        type: integer
      Last stats reset:
        type: string
      Max auth time:
        type: string
      Max session time:
        type: string
      RX:
        type: string
      Sessions handled:
        type: integer
      TX:
        type: string
      Timed out (idle) sessions:
        type: integer
      Timed out sessions:
        type: integer
      raw_avg_auth_time:
        type: integer
      raw_avg_session_time:
        type: integer
      raw_last_stats_reset:
        type: integer
      raw_max_auth_time:
        type: integer
      raw_max_session_time:
        type: integer
      raw_rx:
        description: raw fields if you want
        type: integer
      raw_tx:
        type: integer
    type: object
  home.DockerService:
    properties:
      log_stream:
        $ref: '#/definitions/home.DockerStats'
      ocserv:
        $ref: '#/definitions/home.DockerStats'
      postgres:
        $ref: '#/definitions/home.DockerStats'
      user_expiry:
        $ref: '#/definitions/home.DockerStats'
      web:
        $ref: '#/definitions/home.DockerStats'
    required:
    - log_stream
    - ocserv
    - postgres
    - user_expiry
    - web
    type: object
  home.DockerStats:
    properties:
      cpu:
        $ref: '#/definitions/home.CPU'
      name:
        type: string
      ram:
        $ref: '#/definitions/home.RAM'
    required:
    - name
    type: object
  home.GeneralInfo:
    properties:
      _Up since:
        type: string
      Active sessions:
        type: integer
      IPs in ban list:
        type: integer
      Median latency:
        type: string
      STDEV latency:
        type: string
      Sec-mod PID:
        type: integer
      Sec-mod instance count:
        type: integer
      Server PID:
        type: integer
      Status:
        type: string
      Total authentication failures:
        type: integer
      Total sessions:
        type: integer
      Up since:
        type: string
      raw_median_latency:
        description: raw fields if you want
        type: integer
      raw_stdev_latency:
        type: integer
      raw_up_since:
        type: integer
      uptime:
        type: integer
    type: object
  home.GetHomeResponse:
    properties:
      ip_bans:
        items:
          $ref: '#/definitions/models.IPBanPoints'
        type: array
      statistics:
        items:
          $ref: '#/definitions/models.DailyTraffic'
        type: array
      top_bandwidth_user:
        $ref: '#/definitions/repository.TopBandwidthUsers'
      total_bandwidth:
        $ref: '#/definitions/repository.TotalBandwidths'
      users:
        $ref: '#/definitions/home.GetHomeUser'
    type: object
  home.GetHomeUser:
    properties:
      online_users_session:
        items:
          $ref: '#/definitions/models.OnlineUserSession'
        type: array
      total:
        type: integer
    type: object
  home.OcservStatusResponse:
    properties:
      current_stats:
        $ref: '#/definitions/home.CurrentStats'
      general_info:
        $ref: '#/definitions/home.GeneralInfo'
    type: object
  home.RAM:
    properties:
      total:
        type: number
      used:
        type: number
      used_percent:
        type: number
    type: object
  home.ServerStatusResponse:
    properties:
      cpu:
        $ref: '#/definitions/home.CPU'
      ram:
        $ref: '#/definitions/home.RAM'
      swap:
        $ref: '#/definitions/home.Swap'
    type: object
  home.Swap:
    properties:
      total:
        type: number
      used:
        type: number
      used_percent:
        type: number
    type: object
...
  middlewares.Unauthorized:
    properties:
      error:
        type: string
    type: object
  models.DailyTraffic:
    properties:
      date:
        description: 'Format: YYYY-MM-DD'
        type: string
      rx:
        description: in GiB
        type: number
      tx:
        description: in GiB
        type: number
    type: object
  models.IPBanPoints:
    properties:
      _Since:
        type: string
      IP:
        type: string
      Score:
        type: integer
      Since:
        type: string
    type: object
...
  models.OcservInfo:
    properties:
      status:
        type: string
      version:
        $ref: '#/definitions/models.ServerVersion'
    required:
    - status
    - version
    type: object
  models.OcservUser:
    properties:
      config:
        $ref: '#/definitions/models.OcservUserConfig'
      created_at:
        type: string
      deactivated_at:
        type: string
      description:
        type: string
      expire_at:
        type: string
      group:
        type: string
      is_locked:
        type: boolean
      is_online:
        type: boolean
      owner:
        type: string
      password:
        type: string
      rx:
        description: Receive in bytes
        type: integer
      traffic_size:
        description: in GiB  >> x * 1024 ** 3
        type: integer
      traffic_type:
        enum:
        - Free
        - MonthlyTransmit
        - MonthlyReceive
        - TotallyTransmit
        - TotallyReceive
        type: string
      tx:
        description: Transmit in bytes
        type: integer
      uid:
        type: string
      updated_at:
        type: string
      username:
        type: string
    required:
    - created_at
    - group
    - is_locked
    - is_online
    - owner
    - password
    - rx
    - traffic_size
    - traffic_type
    - tx
    - uid
    - username
    type: object
  models.OcservUserConfig:
    properties:
      dns:
        description: 'Comma-separated list of DNS servers to assign to the user. Example:
          ''8.8.8.8,1.1.1.1'''
        items:
          type: string
        type: array
      explicit-ipv4:
        description: 'Static IPv4 address to assign to the user. Example: ''192.168.100.10'''
        type: string
      idle-timeout:
        description: 'Time in seconds before disconnecting idle users. Example: 600'
        type: integer
      ipv4-network:
        description: 'The pool of addresses from which to assign to the user. Example:
          ''192.168.1.0/24'''
        type: string
      iroute:
        description: 'Internal route available only via VPN. Example: ''10.0.0.0/8'''
        type: string
      mobile-idle-timeout:
        description: 'Idle timeout in seconds for mobile users. Example: 900'
        type: integer
      nbns:
        description: 'NetBIOS Name Servers (WINS) for Windows clients. Example: ''192.168.1.1'''
        type: string
      no-route:
        description: 'List of networks to exclude from VPN routing. Example: [''192.168.0.0/16'',
          ''10.0.0.0/8'']'
        items:
          type: string
        type: array
      rekey-time:
        description: 'Rekey time in seconds; triggers key renegotiation. Example:
          86400 for 24 hours'
        type: integer
      restrict-to-ports:
        description: 'Comma-separated list of allowed or blocked ports/protocols.
          Supports ''tcp(port)'', ''udp(port)'', ''icmp()'', ''icmpv6()'', and negation
          with ''!()''. Example: ''tcp(443), udp(53)'' or ''!(tcp(22), udp(1194))'''
        type: string
      restrict-to-routes:
        description: 'Allow user access only to defined routes. Example: true'
        type: boolean
      route:
        description: 'Routes pushed to the user for routing traffic. Example: [''0.0.0.0/0'',
          ''10.10.0.0/16'']'
        items:
          type: string
        type: array
      session-timeout:
        description: 'Maximum session time in seconds before forced disconnect. Example:
          3600'
        type: integer
      split-dns:
        description: 'List of domains over which the provided DNS servers should be
          used. Example: [''example.com'', ''internal.company.com'']'
        items:
          type: string
        type: array
    type: object
...
  models.OnlineUserSession:
    properties:
      _Connected at:
        type: string
      Average RX:
        type: string
      Average TX:
        type: string
      Groupname:
        type: string
      Username:
        type: string
    type: object
  models.ServerVersion:
    properties:
      occtl_version:
        type: string
      ocserv_version:
        type: string
    type: object
...
  repository.TopBandwidthUsers:
    properties:
      top_rx:
        items:
          $ref: '#/definitions/models.OcservUser'
        type: array
      top_tx:
        items:
          $ref: '#/definitions/models.OcservUser'
        type: array
    type: object
  repository.TotalBandwidths:
    properties:
      rx:
        type: number
      tx:
        type: number
    required:
    - rx
    - tx
    type: object
  request.ErrorResponse:
    properties:
      error:
        items:
          type: string
        type: array
      message:
        items:
          type: string
        type: array
    required:
    - error
    - message
    type: object
...
info:
  contact: {}
  description: This is a sample Ocserv User management Api server.
  title: Ocserv User management Example Api
  version: "1.0"
paths:
...
  /home:
    get:
      consumes:
      - application/json
      description: Content of home
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/home.GetHomeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Content of home
      tags:
      - Home
  /home/container-stats:
    get:
      consumes:
      - application/json
      description: Content of docker system usage stats (cpu, ram, swap)
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/home.DockerService'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Content of docker system usage stats
      tags:
      - Home
  /home/ocserv-stats:
    get:
      consumes:
      - application/json
      description: Content of ocserv server stats
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/home.OcservStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Content of ocserv server stats
      tags:
      - Home
  /home/system-stats:
    get:
      consumes:
      - application/json
      description: Content of os system usage stats (cpu, ram, swap)
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/home.ServerStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Content of os system usage stats
      tags:
      - Home
  /occtl/commands:
    get:
      consumes:
      - application/json
      description: Occtl Commands
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Command Action ID (1 to 15)
        in: query
        name: action
        required: true
        type: integer
      - description: Optional parameter depending on command
        in: query
        name: value
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Occtl Commands
      tags:
      - OCCTL
  /occtl/server_info:
    get:
      consumes:
      - application/json
      description: Server information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OcservInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
      summary: Server information
      tags:
      - OCCTL
...
swagger: "2.0"